refactor(api): move DPF support into machine state controller
Relocate the DPF operations abstraction and SDK wrapper from the API root into
the machine state controller module.

This keeps DPF provisioning, watcher integration, and test mocks colocated with
the machine state handler, reducing API crate coupling as preparation for
extracting the machine state controller into its own crate.
