+++ crates/api/src/lib.rs
@@ -41,7 +41,6 @@ mod credentials;
 mod db_init;
 mod dhcp;
 mod dpa;
-mod dpf;
 mod dpf_services;
 mod dynamic_settings;
 mod errors;
