--- crates/api/src/state_controller/external_service_error.rs
@@ -1,24 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
- * SPDX-License-Identifier: Apache-2.0
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-use carbide_dpf::DpfError;
-
-use crate::state_controller::state_handler::{ExternalServiceError, StateHandlerError};
-
-pub(crate) fn dpf_error(error: DpfError) -> StateHandlerError {
-    ExternalServiceError::with_source("dpf", "", error.to_string(), "dpf_error", error).into()
-}
