+++ crates/api/src/state_controller/machine/mod.rs
@@ -32,6 +32,7 @@ use super::state_handler::StateHandlerError;
 
 pub mod config;
 pub mod context;
+pub mod dpf;
 pub mod handler;
 pub mod io;
 pub mod metrics;
