+++ crates/api/src/state_controller/mod.rs
@@ -16,7 +16,6 @@
  */
 
 pub mod common_services;
-pub(crate) mod external_service_error;
 pub mod machine;
 
 pub use ::state_controller::{controller, db_write_batch, state_change_emitter, state_handler};
