+++ crates/api/src/tests/machine_admin_force_delete.rs
@@ -47,6 +47,7 @@ use tonic::Request;
 
 use crate::api::Api;
 use crate::attestation as attest;
+use crate::state_controller::machine::dpf::{DpfOperations, MockDpfOperations};
 use crate::tests::common;
 
 async fn get_partition_status(api: &Api, ib_partition_id: IBPartitionId) -> IbPartitionStatus {
@@ -743,7 +744,7 @@ async fn test_admin_force_delete_with_dpf_uses_bmc_mac(pool: sqlx::PgPool) {
     let captured_calls: Arc<std::sync::Mutex<DpfCallLog>> =
         Arc::new(std::sync::Mutex::new(Vec::new()));
 
-    let mut mock = crate::dpf::MockDpfOperations::new();
+    let mut mock = MockDpfOperations::new();
 
     mock.expect_register_dpu_device().returning(|_| Ok(()));
     mock.expect_register_dpu_node().returning(|_| Ok(()));
@@ -762,7 +763,7 @@ async fn test_admin_force_delete_with_dpf_uses_bmc_mac(pool: sqlx::PgPool) {
             Ok(())
         });
 
-    let dpf_sdk: Arc<dyn crate::dpf::DpfOperations> = Arc::new(mock);
+    let dpf_sdk: Arc<dyn DpfOperations> = Arc::new(mock);
     let mut config = get_config();
     config.dpf = crate::cfg::file::DpfConfig {
         enabled: true,
